Moved common components & styles under common directory
