+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./scss/App.scss";
-import Header from "./components/Header";
+import Header from "./components/common/Header";
 import Home from "./pages/Home";
 import Help from "./pages/Help";
